+++ micc/templates/package-base/{{cookiecutter.project_name}}/tests/test_{{cookiecutter.package_name}}.py
@@ -1,44 +1,30 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
 """Tests for {{ cookiecutter.package_name }} package."""
 
-import os
-import sys
-import pytest
-
 import {{ cookiecutter.package_name }}
 
-# Make sure that the current directory is the project directory.
-# 'make test" and 'pytest' are generally run from the project directory.
-# However, if we run/debug this file in eclipse, we end up in test
-if os.getcwd().endswith('tests'):
-    echo(f"Changing current working directory"
-         f"\n  from '{os.getcwd()}'"
-         f"\n  to   '{os.path.abspath(os.path.join(os.getcwd(),'..'))}'.\n")
-    os.chdir('..')
-# Make sure that we can import the module being tested. When running
-# 'make test" and 'pytest' in the project directory, the current working
-# directory is not automatically added to sys.path.
-if not ('.' in sys.path or os.getcwd() in sys.path):
-    echo(f"Adding '.' to sys.path.\n")
-    sys.path.insert(0, '.')
 
-def test_hello():
+def test_hello_noargs():
     """Test for {{ cookiecutter.package_name }}.hello()."""
     s = {{ cookiecutter.package_name }}.hello()
     assert s=="Hello world"
 
+
+def test_hello_me():
+    """Test for {{ cookiecutter.package_name }}.hello('me')."""
     s = {{ cookiecutter.package_name }}.hello('me')
     assert s=="Hello me"
     
     
 # ==============================================================================
 # The code below is for debugging a particular test in eclipse/pydev.
-# (normally all tests are run with pytest)
+# (otherwise all tests are normally run with pytest)
+# Make sure that you run this code with the project directory as CWD, and
+# that the source directory is on the path
 # ==============================================================================
 if __name__ == "__main__":
-    the_test_you_want_to_debug = test_hello
+    the_test_you_want_to_debug = test_hello_noargs()
 
     print("__main__ running", the_test_you_want_to_debug)
     the_test_you_want_to_debug()